se agrego una condicion de que el usuario no se pueda eliminar el mismo, esto aplicaria en el caso de los usuarios admin

+++ recursos_graficos_y_logicos/pantallas_de_la_aplicacion/pantalla_admin_de_vista_general_de_usuarios.py
@@ -412,8 +412,14 @@ class PantallaAdminVistaGeneralUsuarios(QWidget, Ui_VistaGeneralUsuarios):
                     # Buscamos el usuario a partir del id del empleado
                     usuario = usuario_servicio.obtener_usuario_por_empleado_id(empleado[0])
                     
-                    usuario_servicio.eliminar_usuario(usuario[0])
-                    print("El usuario se pudo eliminar correctamente")
+                    if usuario[0] == FuncionSistema.id_usuario:
+                        QMessageBox.warning(self, "Error al eliminar el usuario", f"No te puede eliminar a ti mismo")
+                        return
+                    else:
+                        usuario_servicio.eliminar_usuario(usuario[0])
+                    
+                        print("El usuario se pudo eliminar correctamente")
+                    
         except Exception as e:
             
             QMessageBox.warning(self, "Error al eliminar el usuario", f"{e}")
